Day 6!
Thanks for nothing, Windows. CRLF isn't parsed properly on Linux, leaving the CR because it's not expected. Combined with forgetting substr is size-based and not index-based, I didn't remove the char and got the wrong result... That was a waste of time :] used dos2unix to fix it, but I just threw in an augmented substring function to keep myself sane. After that, day 6 wasn't nearly as complicated as it ended up looking thanks to the file mess